relocate all web:api modules into an specific directory
