+++ logstash-core/lib/logstash/api/lib/app/modules/node.rb
