+++ logstash-core/lib/logstash/api/lib/app/modules/stats.rb
